+++ apps/thundermoon_web/lib/thundermoon_web/live/counter_live.ex
@@ -4,15 +4,14 @@ defmodule ThundermoonWeb.CounterLive do
   import Canada.Can
 
   alias Thundermoon.CounterRealm
-  alias Thundermoon.Counter
 
   alias ThundermoonWeb.CounterView
   alias ThundermoonWeb.Endpoint
 
   def mount(_session, socket) do
     Endpoint.subscribe("counter")
-    :ok = CounterRealm.create()
-    digits = Counter.get_digits()
+    {:ok, _} = CounterRealm.create()
+    digits = CounterRealm.get_digits()
     {:ok, assign(socket, digits: digits)}
   end
 
@@ -21,12 +20,12 @@ defmodule ThundermoonWeb.CounterLive do
   end
 
   def handle_event("inc", value, socket) when value in ["1", "10", "100"] do
-    Counter.inc(value)
+    CounterRealm.inc(value)
     {:noreply, socket}
   end
 
   def handle_event("dec", value, socket) when value in ["1", "10", "100"] do
-    Counter.dec(value)
+    CounterRealm.dec(value)
     {:noreply, socket}
   end
 
@@ -36,7 +35,7 @@ defmodule ThundermoonWeb.CounterLive do
   end
 
   def handle_info("update", socket) do
-    digits = Counter.get_digits()
+    digits = CounterRealm.get_digits()
     {:noreply, assign(socket, digits: digits)}
   end
 end